mcp-examples/weather-mcp-py/pyproject.toml
[tool.poetry]
name = "weather-mcp-py"
version = "0.1.0"
description = "Weather MCP Server with Nevermined Payments Integration (Python)"
authors = ["Nevermined"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
payments-py = "^1.0.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
